requirements.txt
PyYAML==6.*
matplotlib==3.*
pytest==8.*
